backend/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
django-filter==21.1
gunicorn==20.0.4
pytz==2020.1
asgiref==3.2.10
psycopg2-binary==2.8.6
sqlparse==0.3.1
python-dotenv==0.20.0
Pillow==9.2.0
django-colorfield==0.7.2
drf_extra_fields==3.4.0
djoser==2.1.0
